MAINT: Keep TAP implementation colocated
